feat: 글쓰기 등 어드민 권한 체크 추가

+++ src/main/java/com/wonkwang/health/controller/PostController.java
@@ -1,5 +1,6 @@
 package com.wonkwang.health.controller;
 
+import com.wonkwang.health.domain.Role;
 import com.wonkwang.health.dto.PostDTO;
 import com.wonkwang.health.dto.ResponseDTO;
 import com.wonkwang.health.dto.ResponseEntityBuilder;
@@ -18,6 +19,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
+import static com.wonkwang.health.dto.ResponseEntityBuilder.*;
 import static org.springframework.http.HttpStatus.*;
 
 @RestController
@@ -31,43 +33,55 @@ public class PostController {
     @GetMapping("/{postId}")
     public ResponseEntity<ResponseDTO<PostDTO>> getPost(@PathVariable Long postId) {
 
-        return ResponseEntityBuilder.build("글 불러오기 완료", OK, postService.getOnePost(postId));
+        return build("글 불러오기 완료", OK, postService.getOnePost(postId));
     }
 
     @GetMapping
     public ResponseEntity<ResponseDTO<Page<PostDTO>>> getPostList(@PageableDefault(size=30, sort = "createdDate", direction = Sort.Direction.DESC) Pageable pageable) {
 
-        return ResponseEntityBuilder.build("글목록 불러오기 완료", OK, postService.getPostList(pageable));
+        return build("글목록 불러오기 완료", OK, postService.getPostList(pageable));
     }
 
     @PostMapping
-    public ResponseEntity<ResponseDTO<Long>> createPost(@RequestBody PostDTO postDTO, @SessionAttribute Long userId) {
-        System.out.println("postDTO = " + postDTO);
+    public ResponseEntity<ResponseDTO<Long>> createPost(@RequestBody PostDTO postDTO, @SessionAttribute Role role) {
+        if (role != Role.ADMIN) {
+            return build("권한 부족", FORBIDDEN, null);
+        }
         Long createdPostId = postService.createPost(postDTO);
-        return ResponseEntityBuilder.build("글 생성 완료", OK, createdPostId);
+        return build("글 생성 완료", OK, createdPostId);
     }
 
     @PatchMapping("/{postId}")
-    public ResponseEntity<ResponseDTO<Long>> updatePost(@RequestBody PostDTO postDTO, @PathVariable Long postId) {
+    public ResponseEntity<ResponseDTO<Long>> updatePost(@RequestBody PostDTO postDTO, @PathVariable Long postId, @SessionAttribute Role role) {
+        if (role != Role.ADMIN) {
+            return build("권한 부족", FORBIDDEN, null);
+        }
         System.out.println("postDTO = " + postDTO);
         postService.updatePost(postId, postDTO);
-        return ResponseEntityBuilder.build("글 수정 완료", OK, postId);
+        return build("글 수정 완료", OK, postId);
     }
 
     @DeleteMapping("/{postId}")
-    public ResponseEntity<ResponseDTO<Long>> deletePost(@PathVariable Long postId) {
+    public ResponseEntity<ResponseDTO<Long>> deletePost(@PathVariable Long postId, @SessionAttribute Role role) {
+        if (role != Role.ADMIN) {
+            return build("권한 부족", FORBIDDEN, null);
+        }
         postService.deleteOnePost(postId);
-        return ResponseEntityBuilder.build("글 삭제 완료", OK, postId);
+        return build("글 삭제 완료", OK, postId);
     }
 
     @PostMapping("/image")
-    public List<String> uploadImages(@RequestParam("file") MultipartFile[] files) throws IOException {
+    public ResponseEntity<ResponseDTO<List<String>>> uploadImages(@RequestParam("file") MultipartFile[] files, @SessionAttribute Role role) throws IOException {
+        if (role != Role.ADMIN) {
+            return build("권한 부족", FORBIDDEN, null);
+        }
+
         List<String> imageUrls = new ArrayList<>();
         for (MultipartFile file : files) {
             String url = s3Service.uploadFile(file);
             imageUrls.add(url);
         }
         System.out.println("Uploaded files: " + imageUrls.size()); // 업로드된 파일 수 출력
-        return imageUrls;
+        return build("이미지 업로드 성공",OK, imageUrls);
     }
 }
